Moved CSS to its own folder
